shell.nix: allow to use different clang-format and clang-tidy versions.
For `clang-format`, we'd like to be a bit more conservative and use
a mature version that does not flip-flop on llvm versions.

For `clang-tidy` we generally want to be at the latest version. For
now, keep it at the same version as before (18) as we first need to
address the new findings.